backend/requirements.txt
fastapi>=0.133.0
uvicorn[standard]>=0.41.0
python-dotenv>=1.2.1
chromadb>=1.5.1
sentence-transformers>=5.2.3
langchain>=1.2.10
langchain-openai>=1.1.10
openai>=2.24.0
langgraph>=1.0.9
ddgs>=9.10.0
pydantic>=2.10.0
